+++ code_king/백준/Silver/9934. 완전 이진 트리/9934_my_output.txt
